Move stray command defs from GitModule to GitCommandHelpers
