+++ UnitTests/GitCommands.Tests/GitModuleTests.cs
@@ -11,7 +11,6 @@ using GitCommands.Git;
 using GitExtUtils;
 using GitUI;
 using GitUIPluginInterfaces;
-using JetBrains.Annotations;
 using NUnit.Framework;
 
 namespace GitCommandsTests
@@ -732,32 +731,6 @@ namespace GitCommandsTests
         public void GetStashesCmd(bool noLocks, string expected)
         {
             Assert.AreEqual(expected, _gitModule.GetStashesCmd(noLocks).ToString());
-        }
-
-        [TestCase(@"-c diff.submodule=short -c diff.noprefix=false -c diff.mnemonicprefix=false -c diff.ignoreSubmodules=none diff --no-color -M -C --cached extra -- ""new"" ""old""", "new", "old", true, "extra", false)]
-        [TestCase(@"-c diff.submodule=short -c diff.noprefix=false -c diff.mnemonicprefix=false -c diff.ignoreSubmodules=none diff --no-color extra -- ""new""", "new", "old", false, "extra", false)]
-        [TestCase(@"--no-optional-locks -c diff.submodule=short -c diff.noprefix=false -c diff.mnemonicprefix=false -c diff.ignoreSubmodules=none diff --no-color -M -C --cached extra -- ""new"" ""old""", "new", "old", true, "extra", true)]
-        public void GetCurrentChangesCmd(string expected, string fileName, [CanBeNull] string oldFileName, bool staged,
-            string extraDiffArguments, bool noLocks)
-        {
-            Assert.AreEqual(expected, _gitModule.GetCurrentChangesCmd(fileName, oldFileName, staged,
-                extraDiffArguments, noLocks).ToString());
-        }
-
-        [TestCase(@"for-each-ref --sort=-committerdate --format=""%(objectname) %(refname)"" refs/heads/", false)]
-        [TestCase(@"--no-optional-locks for-each-ref --sort=-committerdate --format=""%(objectname) %(refname)"" refs/heads/", true)]
-        public void GetSortedRefsCommand(string expected, bool noLocks)
-        {
-            Assert.AreEqual(expected, _gitModule.GetSortedRefsCommand(noLocks).ToString());
-        }
-
-        [TestCase(@"show-ref --dereference", true, true, false)]
-        [TestCase(@"show-ref --tags", true, false, false)]
-        [TestCase(@"for-each-ref --sort=-committerdate refs/heads/ --format=""%(objectname) %(refname)""", false, true, false)]
-        [TestCase(@"--no-optional-locks for-each-ref --sort=-committerdate refs/heads/ --format=""%(objectname) %(refname)""", false, true, true)]
-        public void GetRefsCmd(string expected, bool tags, bool branches, bool noLocks)
-        {
-            Assert.AreEqual(expected, _gitModule.GetRefsCmd(tags, branches, noLocks).ToString());
         }
 
         [TestCase(false, @"-c core.safecrlf=false update-index --add --stdin")]
